requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-extensions==4.1
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
django-environ==0.12.0
psycopg==3.2.9
psycopg-binary==3.2.9
django-cors-headers==4.7.0
gunicorn==23.0.0
celery==5.5.2
redis==6.1.0
ruff==0.11.11